requirements.txt
autopep8==2.0.0
numpy==1.23.5
pycodestyle==2.10.0
PyOpenGL==3.1.6
PyQt6==6.4.0
PyQt6-Qt6==6.4.1
PyQt6-sip==13.4.0
scipy==1.9.3
tomli==2.0.1
trimesh==3.17.1
